Fix infinite loop when splitting a transcript into chunks
I asked ChatGPT for help with this one because I'm bad at
off-by-one errors and this bug had the distinct smell of an
off-by-one error.

It might be good to add some tests, but that would require
separating the main program from the utility functions,
which is work we better save for a later commit.